crawler/requirements.txt
chromedriver-binary
pandas
Pyrebase4
requests
selenium